feat(zipWith): add zipWith which is just a rename of zip
- moves tests around to split off legacy behaviors
