+++ src/internal/operators/index.ts
@@ -98,5 +98,5 @@ export { windowTime } from './windowTime';
 export { windowToggle } from './windowToggle';
 export { windowWhen } from './windowWhen';
 export { withLatestFrom } from './withLatestFrom';
-export { zip } from './zip';
+export { zip, zipWith } from './zipWith';
 export { zipAll } from './zipAll';
